convert page-skin to TS
